Добавил RRT (#18)
1. Сделал rrt
2. Добавл команду rrt
3. Добавил общий генератор случайных чисел с режимами в конфиг
4. Немного рефакторинга
5. Добавил непрерывные методы
6. Улучшил безопасность config
7. Починил визуализацию gnuplot + цветной bmp
8. Убрал ненужные режимы bin
9. Протестировал корректность программы
10. Добавил строгие флаги компилятору и линковщику. Программу исправил, чтобы компилировалась
11. Добавил визуализацию RRT
12. Обновил датасеты в исследованиях, после исправлений в коде
13. Новое исследование RRT

+++ src/algorithms/geometry/voronoi_diagram.cpp
@@ -83,8 +83,8 @@ namespace algorithms::geometry {
         return;
     }
     
-    const int width = field[0].size();
-    const int height = field.size();
+    const int width = static_cast<int>(field[0].size());
+    const int height = static_cast<int>(field.size());
     edges.clear();
 
     logger.debug(std::string("[VoronoiDiagram::buildFromDelaunay] Размер области: ") +
